Moved `Colored` to its own module
